% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_calculate_mode}
\alias{arlc_calculate_mode}
\title{Calculate the Mode of a Vector}
\usage{
arlc_calculate_mode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The mode of the vector.
}
\description{
This function calculates the mode of a vector.
}
\examples{
arlc_calculate_mode(c(1, 2, 2, 3, 4))
}
