% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_clean_final_rules.R
\name{arlc_clean_final_rules}
\alias{arlc_clean_final_rules}
\title{Clean Final Rules}
\usage{
arlc_clean_final_rules(final_rules)
}
\arguments{
\item{final_rules}{A set of final rules to be cleaned.}
}
\value{
A cleaned set of rules.
}
\description{
This function cleans the final set of association rules.
}
\examples{
\dontrun{
library(arules)
# Create a sample transactions dataset
data("Adult")
transactions <- as(Adult, "transactions")
# Generate rules
rules <- apriori(transactions, parameter = list(supp = 0.5, conf = 0.9))
# Clean the final rules
cleaned_rules <- arlc_clean_final_rules(rules)
}
}
