% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_clean_transactions.R
\name{arlc_clean_transactions}
\alias{arlc_clean_transactions}
\title{Clean Transactional Datasets}
\usage{
arlc_clean_transactions(all_sets, file, genFileLog = FALSE, debugMode = FALSE)
}
\arguments{
\item{all_sets}{A list of transaction sets.}

\item{file}{A string representing the file name used for logging.}

\item{genFileLog}{A boolean flag indicating whether to generate a log file. Default is FALSE.}

\item{debugMode}{A boolean flag indicating whether to print debug information. Default is FALSE.}
}
\value{
A list of filtered transaction sets.
}
\description{
This function filters out fully overlapping sets from a list of transaction sets.
}
\examples{
arlc_clean_transactions(list(c(1,2,3), c(2,3), c(1,2,3,4)), "example_file", TRUE, TRUE)
}
