% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_clusters_plot.R
\name{arlc_clusters_plot}
\alias{arlc_clusters_plot}
\title{Plot Graph with Custom Layout and Communities}
\usage{
arlc_clusters_plot(g, graphLabel, clusters)
}
\arguments{
\item{g}{An igraph object representing the graph.}

\item{graphLabel}{A character string for the graph label to be displayed in the title.}

\item{clusters}{A list of clusters to highlight in the plot.}
}
\value{
The function produces a plot as a side effect.
}
\description{
This function plots a graph with specified aesthetics and highlights communities.
}
\examples{
\dontrun{
library(igraph)
g <- make_ring(10)
clusters <- list(1:5, 6:10)
arlc_clusters_plot(g, "Sample Graph", clusters)
}
}
