% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_convert_date_format}
\alias{arlc_convert_date_format}
\title{Convert a Date to a Different Format}
\usage{
arlc_convert_date_format(date_str, from_format, to_format)
}
\arguments{
\item{date_str}{A date string.}

\item{from_format}{The current format of the date string.}

\item{to_format}{The desired format of the date string.}
}
\value{
The date string in the new format.
}
\description{
This function converts a date from one format to another.
}
\examples{
arlc_convert_date_format("2023-01-01", "\%Y-\%m-\%d", "\%d-\%m-\%Y")
}
