% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_count_na}
\alias{arlc_count_na}
\title{Count NA Values in a Data Frame}
\usage{
arlc_count_na(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A named vector with the count of NA values in each column.
}
\description{
This function counts the number of NA values in each column of a data frame.
}
\examples{
arlc_count_na(data.frame(a = c(1, NA, 3), b = c(NA, NA, 3)))
}
