% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_df_summary}
\alias{arlc_df_summary}
\title{Create a Summary of a Data Frame}
\usage{
arlc_df_summary(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame summarizing the statistics of the input data frame.
}
\description{
This function creates a summary of a data frame including count, mean, median, and standard deviation for numeric columns.
}
\examples{
arlc_df_summary(data.frame(a = c(1, 2, 3, 4, 5), b = c(5, 4, 3, 2, 1)))
}
