% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_fct_get_best_apriori_thresholds.R
\name{arlc_fct_get_best_apriori_thresholds}
\alias{arlc_fct_get_best_apriori_thresholds}
\title{Get Best Apriori Thresholds}
\usage{
arlc_fct_get_best_apriori_thresholds(
  transactions,
  support_range,
  confidence_range
)
}
\arguments{
\item{transactions}{A transaction dataset of class \code{transactions} from the \code{arules} package.}

\item{support_range}{A numeric vector specifying the range of support values to be tested.}

\item{confidence_range}{A numeric vector specifying the range of confidence values to be tested.}
}
\value{
A numeric vector containing the best support, best confidence, highest lift, and the number of rules found.
The return value is a named vector with elements \code{best_support}, \code{best_confidence}, \code{best_lift}, and \code{len_rules}.
}
\description{
This function finds the best support and confidence thresholds for the Apriori algorithm
to maximize the lift of the generated association rules.
}
\details{
This function iterates through the given ranges of support and confidence values,
applies the Apriori algorithm to find association rules for each pair of values, and selects
the pair that produces rules with the highest lift.
}
\examples{
\dontrun{
library(arules)
data(Groceries)
support_range <- seq(0.001, 0.01, by = 0.001)
confidence_range <- seq(0.1, 0.9, by = 0.1)
best_thresholds <- arlc_fct_get_best_apriori_thresholds(Groceries, support_range, confidence_range)
print(best_thresholds)
}

}
