% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_file_exists_readable}
\alias{arlc_file_exists_readable}
\title{Check if a File Exists and is Readable}
\usage{
arlc_file_exists_readable(filepath)
}
\arguments{
\item{filepath}{The path to the file.}
}
\value{
TRUE if the file exists and is readable, FALSE otherwise.
}
\description{
This function checks if a file exists and is readable.
}
\examples{
arlc_file_exists_readable("example.txt")
}
