% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_gen_gross_rules.R
\name{arlc_gen_gross_rules}
\alias{arlc_gen_gross_rules}
\title{Get Gross Rules}
\usage{
arlc_gen_gross_rules(trx, minSupp, minConf, minLenRules, maxLenRules)
}
\arguments{
\item{trx}{A transaction dataset.}

\item{minSupp}{Minimum support threshold for the Apriori algorithm.}

\item{minConf}{Minimum confidence threshold for the Apriori algorithm.}

\item{minLenRules}{Minimum length of the rules to be generated.}

\item{maxLenRules}{Maximum length of the rules to be generated.}
}
\value{
A list containing the total number of gross rules with length filter and the gross rules themselves.
}
\description{
This function takes a transaction dataset and minimum support, confidence, minimum length, and maximum length of rules,
and executes the Apriori algorithm to generate gross rules.
}
\details{
This function executes the Apriori algorithm to generate gross rules based on the given parameters.
}
\examples{
arlc_gen_gross_rules(transactions, minSupp = 0.1, minConf = 0.5, minLenRules = 1, maxLenRules = 3)
}
