% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_gen_transactions.R
\name{arlc_gen_transactions}
\alias{arlc_gen_transactions}
\title{Get Transactional Dataset}
\usage{
arlc_gen_transactions(graph)
}
\arguments{
\item{graph}{A graph object which could be loaded via the following call:
g <- arlc_get_network_dataset("./data/karate.gml", "Karate Club")}
}
\value{
A transactional dataset.
}
\description{
This function takes a graph dataset as input, extracts its characteristics,
converts it to an adjacency matrix, and then to transactional dataset.
It filters out transactions with only one element and returns the resulting dataset.
}
\details{
This function generates a transactional dataset from a graph dataset.
}
\examples{
trans <- arlc_gen_transactions(g$graph)

}
