% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_generate_date_sequence}
\alias{arlc_generate_date_sequence}
\title{Generate a Sequence of Dates}
\usage{
arlc_generate_date_sequence(start_date, end_date, by = "day")
}
\arguments{
\item{start_date}{The start date.}

\item{end_date}{The end date.}

\item{by}{The step size for the sequence (e.g., "day", "week", "month").}
}
\value{
A vector of dates.
}
\description{
This function generates a sequence of dates between two given dates.
}
\examples{
arlc_generate_date_sequence("2023-01-01", "2023-01-10", "day")
}
