% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_generate_uid}
\alias{arlc_generate_uid}
\title{Generate a Unique Identifier}
\usage{
arlc_generate_uid(length = 10)
}
\arguments{
\item{length}{The length of the unique identifier. Default is 10.}
}
\value{
A unique identifier string.
}
\description{
This function generates a unique identifier string.
}
\examples{
arlc_generate_uid()
arlc_generate_uid(15)
}
