% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_get_apriori_thresholds.R
\name{arlc_get_apriori_thresholds}
\alias{arlc_get_apriori_thresholds}
\title{Get Apriori Thresholds}
\usage{
arlc_get_apriori_thresholds(trx, supportRange, confidenceRange)
}
\arguments{
\item{trx}{A transaction dataset.}

\item{supportRange}{A sequence of values representing the range for minimum support.}

\item{confidenceRange}{A sequence of values representing the range for minimum confidence.}
}
\value{
A list containing the best minimum support, minimum confidence, best lift,
total number of gross rules, and ratio of generated rules to total number of transactions.
}
\description{
This function takes a transaction dataset and ranges for support and confidence,
computes the best thresholds, and returns the best minimum support, minimum confidence,
best lift, total number of gross rules, and ratio of generated rules to total number of transactions.
}
\details{
This function generates gross rules based on the best obtained thresholds.
}
\examples{
arlc_get_apriori_thresholds(trx, supportRange = seq(0.1, 0.9, by = 0.1), confidenceRange = seq(0.5, 0.9, by = 0.1))
}
