% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_get_best_apriori_thresholds.R
\name{arlc_get_best_apriori_thresholds}
\alias{arlc_get_best_apriori_thresholds}
\title{Get Best Apriori Thresholds}
\usage{
arlc_get_best_apriori_thresholds(transactions, support_range, confidence_range)
}
\arguments{
\item{transactions}{A transaction dataset.}

\item{support_range}{A sequence of values representing the range for minimum support.}

\item{confidence_range}{A sequence of values representing the range for minimum confidence.}
}
\value{
A numeric vector containing the best support, confidence, lift, and number of rules found.
}
\description{
This function takes a transaction dataset and ranges for support and confidence,
runs the Apriori algorithm with various combinations of support and confidence values,
and returns the best support, confidence, lift, and number of rules found.
}
\details{
This function determines the best Apriori thresholds for a given transaction dataset.
}
\examples{
arlc_get_best_apriori_thresholds(transactions, support_range = seq(0.1, 0.9, by = 0.1),
                    confidence_range = seq(0.5, 0.9, by = 0.1))
}
