% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_is_numeric_vector}
\alias{arlc_is_numeric_vector}
\title{Check if a Vector is Numeric}
\usage{
arlc_is_numeric_vector(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
TRUE if all elements are numeric, FALSE otherwise.
}
\description{
This function checks if all elements of a vector are numeric.
}
\examples{
arlc_is_numeric_vector(c(1, 2, 3))
arlc_is_numeric_vector(c(1, "a", 3))
}
