% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_list_to_df}
\alias{arlc_list_to_df}
\title{Convert List of Vectors to Data Frame}
\usage{
arlc_list_to_df(lst)
}
\arguments{
\item{lst}{A list of named vectors.}
}
\value{
A data frame with each element of the list as a row.
}
\description{
This function converts a list of named vectors to a data frame.
}
\examples{
lst <- list(a = c(x = 1, y = 2), b = c(x = 3, y = 4))
arlc_list_to_df(lst)
}
