% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_log_message}
\alias{arlc_log_message}
\title{Log Messages to a File}
\usage{
arlc_log_message(message, file, append = TRUE)
}
\arguments{
\item{message}{The message to log.}

\item{file}{The file where the message should be logged.}

\item{append}{Logical, whether to append to the file or overwrite it. Default is TRUE.}
}
\value{
None
}
\description{
This function logs messages to a specified file.
}
\examples{
arlc_log_message("This is a test message", "log.txt")
}
