% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_measure_time}
\alias{arlc_measure_time}
\title{Measure Execution Time of a Function}
\usage{
arlc_measure_time(func, ...)
}
\arguments{
\item{func}{The function to measure.}

\item{...}{Additional arguments to pass to the function.}
}
\value{
The result of the function execution.
}
\description{
This function measures the execution time of a given function.
}
\examples{
arlc_measure_time(Sys.sleep, 1)
}
