% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_normalize_vector}
\alias{arlc_normalize_vector}
\title{Normalize a Numeric Vector}
\usage{
arlc_normalize_vector(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A normalized numeric vector.
}
\description{
This function normalizes a numeric vector to have values between 0 and 1.
}
\examples{
arlc_normalize_vector(c(1, 2, 3, 4, 5))
}
