% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_utils.R
\name{arlc_replace_na}
\alias{arlc_replace_na}
\title{Replace NA with a Specified Value}
\usage{
arlc_replace_na(x, value)
}
\arguments{
\item{x}{A vector or data frame.}

\item{value}{The value to replace NA with.}
}
\value{
The vector or data frame with NA values replaced.
}
\description{
This function replaces NA values in a vector or data frame with a specified value.
}
\examples{
arlc_replace_na(c(1, NA, 3), 0)
arlc_replace_na(data.frame(a = c(1, NA, 3), b = c(NA, NA, 3)), 0)
}
