% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_rules.R
\name{get_sig_rules}
\alias{get_sig_rules}
\title{Get Significant Rules}
\usage{
get_sig_rules(all_trans, nonRR_rules, method = "fisher", adjust = "bonferroni")
}
\arguments{
\item{all_trans}{A dataframe containing all transactions.}

\item{nonRR_rules}{A list of non-redundant rules.}

\item{method}{The method used to determine significance (e.g., "fisher").}

\item{adjust}{The adjustment method for p-values (e.g., "bonferroni").}
}
\value{
A list containing the total number of significant non-redundant rules and the significant rules themselves.
}
\description{
This function takes all transactions and a set of non-redundant rules as input,
and returns significant rules based on a specified method and adjustment.
}
\details{
This function filters significant rules from a set of non-redundant rules.
}
\examples{
get_sig_rules(all_transactions, non_redundant_rules, method = "fisher", adjust = "bonferroni")
}
