% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trans.R
\name{get_trans}
\alias{get_trans}
\title{Get Transactional Dataset}
\usage{
get_trans(graph)
}
\arguments{
\item{graph}{A graph object.}
}
\value{
A transactional dataset.
}
\description{
This function takes a graph object as input, extracts its characteristics,
converts it to an adjacency matrix, and then to transactions.
It filters out transactions with only one element and returns the resulting dataset.
}
\details{
This function generates a transactional dataset from a graph.
}
\examples{
get_trans(graph)
}
