% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_clean_final_rules.R
\name{arlc_clean_final_rules}
\alias{arlc_clean_final_rules}
\title{Clean Final Rules}
\usage{
arlc_clean_final_rules(final_rules)
}
\arguments{
\item{final_rules}{A dataframe containing the preprocessed rules.}
}
\value{
A cleaned vector of rules.
}
\description{
This function takes a set of preprocessed rules, removes redundant
and significant rules, and returns a cleaned vector of rules.
}
\details{
Cleans the preprocessed rules by removing redundant and significant rules.
}
\examples{
arlc_clean_final_rules(final_rules)
}
