% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_gen_gross_rules.R
\name{arlc_gen_gross_rules}
\alias{arlc_gen_gross_rules}
\title{Get Gross Rules}
\usage{
arlc_gen_gross_rules(transactions, minSupp, minConf, minLenRules, maxLenRules)
}
\arguments{
\item{transactions}{A transactions object.}

\item{minSupp}{Minimum support threshold.}

\item{minConf}{Minimum confidence threshold.}

\item{minLenRules}{Minimum length of rules.}

\item{maxLenRules}{Maximum length of rules.}
}
\value{
A set of gross association rules.
}
\description{
This function generates gross association rules from transactions.
}
\examples{
\dontrun{
library(arules)
# Create a sample transactions dataset
data("Adult")
transactions <- as(Adult, "transactions")
# Generate gross rules
arlc_gen_gross_rules(transactions, minSupp = 0.1, minConf = 0.5, minLenRules = 1, maxLenRules = 3)
}
}
