% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_gen_transactions.R
\name{arlc_gen_transactions}
\alias{arlc_gen_transactions}
\title{Get Transactional Dataset}
\usage{
arlc_gen_transactions(graph)
}
\arguments{
\item{graph}{A graph object.}
}
\value{
A transactional dataset.
}
\description{
This function generates a transactional dataset from a graph.
}
\examples{
\dontrun{
library(igraph)
# Create a sample graph
g <- make_ring(10)
# Generate transactional dataset
trans <- arlc_gen_transactions(g)
}
}
