% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_generate_clusters.R
\name{arlc_generate_clusters}
\alias{arlc_generate_clusters}
\title{Generate Clusters}
\usage{
arlc_generate_clusters(vec)
}
\arguments{
\item{vec}{A vector of preprocessed rules.}

\item{minSupp}{The minimum support threshold for generating clusters.}

\item{minConf}{The minimum confidence threshold for generating clusters.}
}
\value{
A list of potential clusters.
}
\description{
This function takes a vector of preprocessed rules, combines elements
starting with the same value, groups segments by the starting value, sorts elements
within each segment, and returns potential clusters.
}
\details{
This function generates potential clusters based on preprocessed rules.
}
\examples{
arlc_generate_clusters(vec, minSupp = 0.5, minConf = 0.6)
}
