% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_get_network_dataset.R
\name{arlc_get_network_dataset}
\alias{arlc_get_network_dataset}
\title{Get Network Dataset}
\usage{
arlc_get_network_dataset(file_path, label)
}
\arguments{
\item{file_path}{The file path to the GML file to be loaded.}

\item{label}{A label for the graph.}
}
\value{
A list containing the graph object and its properties: total edges, total nodes, and average degree.
}
\description{
This function reads a network dataset from a GML file, assigns node names, and calculates
various properties of the graph such as total edges, total nodes, and average degree.
}
\details{
This function loads a network dataset from a specified GML file and computes basic graph properties.
}
\examples{
\dontrun{
file_path <- system.file("extdata", "karate.gml", package = "arlclustering")
loaded_karate <- arlc_get_network_dataset(file_path, "Karate Club")
print(loaded_karate$graph)
print(loaded_karate$graphLabel)
print(loaded_karate$totalEdges)
print(loaded_karate$graphEdges)
print(loaded_karate$totalNodes)
print(loaded_karate$graphNodes)
print(loaded_karate$averageDegree)
}
}
