% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arlc_get_network_dataset.R
\name{arlc_get_network_dataset}
\alias{arlc_get_network_dataset}
\title{Get Network Dataset}
\usage{
arlc_get_network_dataset(file_path, label)
}
\arguments{
\item{label}{A label for the graph.}

\item{file}{The name of the GML file to be loaded.}
}
\value{
A list containing the graph object and its properties: total edges, total nodes, and average degree.
}
\description{
This function reads a network dataset from a GML file, assigns node names, and calculates
various properties of the graph such as total edges, total nodes, and average degree.
}
\details{
This function loads a network dataset from a specified GML file and computes basic graph properties.
}
\examples{
loaded_karate <- arlc_get_network_dataset("./data/karate.gml", "Karate Club")
loaded_karate
print(loaded_karate$graph)
print(loaded_karate$graphLabel)
print(loaded_karate$totalEdges)
print(loaded_karate$graphEdges)
print(loaded_karate$totalNodes)
print(loaded_karate$graphNodes)
print(loaded_karate$averageDegree)

loaded_dolphins <- arlc_get_network_dataset("./data/dolphins.gml", "Dolphins")
loaded_dolphins
print(loaded_dolphins$graph)
print(loaded_dolphins$graphLabel)
print(loaded_dolphins$totalEdges)
print(loaded_dolphins$graphEdges)
print(loaded_dolphins$totalNodes)
print(loaded_dolphins$graphNodes)
print(loaded_dolphins$averageDegree)

}
