\name{clean_final_rules}
\alias{clean_final_rules}
\title{Clean Final Rules}
\description{
  This function cleans the preprocessed rules to remove redundancy and insignificance.
}
\usage{
  clean_final_rules(final_rules)
}
\arguments{
  \item{final_rules}{The preprocessed rules to clean.}
}
\examples{
  \dontrun{
    cleaned_rules <- clean_final_rules(gross_rules)
  }
}
