% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NonR_rules.R
\name{get_NonR_rules}
\alias{get_NonR_rules}
\title{Get Non-Redundant Rules}
\usage{
get_NonR_rules(gross_rules)
}
\arguments{
\item{gross_rules}{A vector or dataframe of gross rules.}
}
\value{
A list containing the total number of non-redundant rules and the non-redundant rules.
}
\description{
This function takes a set of gross rules, removes redundant rules,
and returns the total number of non-redundant rules along with the non-redundant rules.
}
\details{
This function cleans the gross rules and provides non-redundant rules.
}
\examples{
get_NonR_rules(gross_rules)
}
